프로필 수정 및 이미지 파일 업로드

+++ src/main/java/eclipse/demo/api/BoardApiController.java
@@ -8,10 +8,15 @@ import eclipse.demo.service.BoardService;
 import eclipse.demo.service.MemberService;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.security.Security;
 
 @RestController
 @RequiredArgsConstructor
@@ -22,26 +27,32 @@ public class BoardApiController {
     private final BoardLikeService boardLikeService;
 
     @PostMapping("/boardLike")
-    public String saveLike(@AuthenticationPrincipal Member member, @RequestBody BoardLikeRequest boardLikeRequest) {
+    public String saveLike(@RequestBody BoardLikeRequest boardLikeRequest) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication instanceof AnonymousAuthenticationToken) {
+            throw new ClassCastException("로그인을 해주세요");
+        } else {
+            Member member = (Member) authentication.getPrincipal();
+            Board findBoard = boardService.findOne(boardLikeRequest.getBoardId());
+            Member findMember = memberService.findOne(member.getId());
+            BoardLike findBoardLike = boardLikeService.findLike(findMember.getId(), findBoard.getId());
 
-        Board findBoard = boardService.findOne(boardLikeRequest.getBoardId());
-        Member findMember = memberService.findOne(member.getId());
-        BoardLike findBoardLike = boardLikeService.findLike(findMember.getId(), findBoard.getId());
-
-        // 빈하트일 경우
-        if (boardLikeRequest.getStatus().equals("empty")) {
-            boardLikeService.saveBoardLike(findMember, findBoard);
-        }
-        // 꽉찬하트일 경우
-        else{
-            boardLikeService.delete(findBoardLike.getId());
-        }
+            // 빈하트일 경우
+            if (boardLikeRequest.getStatus().equals("empty")) {
+                boardLikeService.saveBoardLike(findMember, findBoard);
+            }
+            // 꽉찬하트일 경우
+            else {
+                boardLikeService.delete(findBoardLike.getId());
+            }
             return "ok";
+        }
+
     }
 
 
     @Data
-    static class BoardLikeRequest{
+    static class BoardLikeRequest {
         private String status;
         private Long boardId;
     }
